requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.42.1
idna==3.4
imageio==2.31.1
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
networkx==3.0
numpy==1.25.2
onnx==1.14.0
onnxconverter-common==1.14.0
onnxruntime-gpu==1.15.0
opencv-python==4.8.0.76
packaging==23.1
Pillow==9.3.0
protobuf==3.20.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
skl2onnx==1.15.0
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2023.8.12
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
typing_extensions==4.7.1
urllib3==1.26.13
zipp==3.16.2
